requirements.txt
python-telegram-bot
atlassian-python-api
